rename dumb formatter to compat and add a template for it
